Reorganize and refactor files
