+++ src/matchrunner/matchqueue.ts
@@ -1,0 +1,11 @@
+import {Config} from '../config';
+
+export default class MatchQueue {
+
+  // The public div
+  readonly div: HTMLDivElement;
+
+  constructor() {
+
+  }
+}
